Issue: WSW-26261 [RT-1170][Firecrest][RF-Test-Mode] UL-OFDMA (he_tb & trigger_frame) commands are missing in the Wi-Fi Test mode CLI.
[Issue]
Tx_Frame were not getting tramsmitted.

[FIX]
Change the base value of strtol of parameter data_rate.
Change strtol to strtoul for parameter frame_pattern.

[Validation done]
Build successfully without any errors or warnings on armgcc compiler and MCUXpresso IDE.
Basic sanity tests done as uAP and STA using wifi_test_mode app.

[Open Issues/Todo]
None.

[Reviewers]
Abhinav Kulkarni, Sandip Dalvi, Supriya Inamdar, Sanjay Yadav

+++ incl/wifidriver/wifi.h
@@ -1465,7 +1465,7 @@ int wifi_set_rf_tx_frame(const uint32_t enable,
                          const uint32_t data_rate,
                          const uint32_t frame_pattern,
                          const uint32_t frame_length,
-                         const uint32_t adjust_burst_sifs,
+                         const uint16_t adjust_burst_sifs,
                          const uint32_t burst_sifs_in_us,
                          const uint32_t short_preamble,
                          const uint32_t act_sub_ch,
@@ -1474,7 +1474,7 @@ int wifi_set_rf_tx_frame(const uint32_t enable,
                          const uint32_t tx_bf,
                          const uint32_t gf_mode,
                          const uint32_t stbc,
-                         const uint32_t *bssid);
+                         const uint8_t *bssid);
 #endif
 #ifdef CONFIG_WIFI_FW_DEBUG
 /** This function registers callbacks which are used to generate FW Dump on USB
